patch

+++ adversarial_attack/adversarial_attack2.py
@@ -968,7 +968,7 @@ def visualize(names, model, device, image_path, numpy_patch_path, offset, height
         #     # print(combined_img)
         #     # img = transform(combined_img[0,:,:,:])
             # img = cv2.cvtColor(np.array(combined_img.int()), cv2.COLOR_BGR2RGB)
-            img  = combined_img[0,:,:,:].int().permute(1, 2, 0).cpu().numpy()
+            img  = combined_img[0,:,:,:].permute(1, 2, 0).cpu().numpy()
             img = T.ToPILImage(mode=None)(img)
             print(img)
